Add functionality to use decklists for desired cards
This will allow the user to use the decks they have built in HDT to determine what cards are desired. This way, instead of just having the desired cards be the cards not in collection, it can be more focused to cards that the user needs for actual decks but does not yet have.