Added filter for only desired cards
I had to rename the labels so that they would be consistent. When the labels changed, if both the missing and desired switches were off they would both say "Show all", and I could not come up with a short way to differentiate them that wouldn't look awkward. I instead made the on and off labels the same so the user knows what setting it is referring to at all times. I renamed the Golden offlabel so that it would be consistent with the other two.